practice/backend/requirements.txt
langchain
langchain-openai
langchain_community
flask
flask-cors
sentence-transformers
faiss-cpu
python-dotenv